+++ examples/simple_parser.cpp
@@ -5,7 +5,7 @@
 
 // This example demonstrates how to parse TOML from a file or stdin and re-serialize it (print it out) to stdout.
 
-#include "examples.hpp"
+#include "examples.h"
 
 #define TOML_ENABLE_UNRELEASED_FEATURES 1
 #include <toml++/toml.h>